Final Version/requirements.facenet.txt
flask
gunicorn
numpy
Pillow
opencv-python
